Make BBQ prompts identical to HELM's version (#39)
In PR #33 I mentioned the only difference was how training examples get sampled. In this PR I'm porting the logic from HELM for how to do sampling.

My goal is to get NewHELM to produce the exact same value for the BBQ stats when using GPT2, as a way to ensure we have a fully functioning replacement.