Update SUTs to use "native + translate" interface. (#52)
The argument for why to structure SUTs this way is in [this doc](https://docs.google.com/document/d/1qzTAaL410fRs8vMNx1lvkewaDM0ASJlzlafR03_6iC4/edit). I figure we can use this code example to help us decide if that doc is a good idea.

This also allows the Prompt to set options that were previously hard coded.