Create a better command line for NewHELM. (#60)
With this change, you can specify which test+sut pair you want to run, or benchmark+sut. Also, this means plugins can add to the command line, so runners can have their own interface!

One kinda gross thing this required was registering Tests, which I'm not convinced are actually enumerable. But that functionality is really only for one-off testing, so should be fine for now.