Rewrite credential management to remove inversion. (#68)
This is trying to resolve the issue @yifanmai pointed out on #55 that the runner is taking on a weird relationship to the SUT. In this version, we can still discover all secrets someone might use, with documentation on how to obtain a value.  It makes credential management more isolated to the objects using it. However, we lose the ability to know where secrets are being used or how they are being set.

To my knowledge existing tooling like Dynaconf doesn't allow for discovery or enforced documentation of values, which I really want.