Coffee uses newhelm (#95)
This does basically the same thing as before, but using NewHelm instead of Helm. Note that NewHelm is much more fragile, so you may have to try any given run a number of times to get it to work, but it does work for me on occasion. Note also that the --web-only option doesn't work for the moment, as we'll have to discuss whether coffee should just save scores or whether it's better to pull things out of the newhelm run data.
---------

Co-authored-by: dhosterman <daniel@danielhosterman.com>