Make get_dependencies and get_metadata be instance methods. (#144)
These had previously been class methods on the idea that you'd want to be able to call them without creating an instance. However, we've since moved to embrace the idea that Tests are instances, for example to support config defined Tests. In that world, it doesn't make sense for all Tests using the same class to use the same dependencies.