Implement the @record_init decorator to write down the init call. (#137)
The idea is that we can use this to recreate a Test/SUT from the TestRecord (see next PR).

Recording the init call is preferable to pickling / etc as objects may have non-in-memory-state (e.g. SUT establishing a remote connection), or arbitrarily large internal state (e.g. a Test that, for some reason, loads a 6 GB file into `self`). Making it a decorator reduces the chance someone will forget to record at all / miss recording an argument.