Add a self-calibration run (#110)
Add a self-calibration run. `run.py calibrate` will show the current calibration, and `run.py --update` will update it.  It uses far too few items for now, but this makes it less likely to crash. Once newhelm is more stable, we should up the number of examples to something where the numbers are stable run to run.