Model chat as a different kind of Prompt available in PromptWithContext (#160)
When a Test makes its `PromptWithContext`, it can now set `prompt` to be either `TextPrompt` or `ChatPrompt`. SUTs are required to implement two incoming translation methods: `translate_text_prompt` and `translate_chat_prompt`. The runner is responsible for calling the right translation, and the expectation is both calls can return the same RequestType.

Some advantages of this design:

* SUTs get to decide how to do the translation, so they can flatten chat / embed a text prompt however they see fit.
* If a SUT really doesn't want to handle of these, it can raise an exception in the `translate_` method.
* Test writing can be pretty natural as you use the object that works for you.

Some disadvantages:

* SUTs have to write two translate request methods.
* For Together, it seems plausible you'd want to switch from TogetherChatSUT to TogetherCompletionsSUT depending on the input, instead of having the model stuck with a specific API. You can achieve this in the current setup, but its kinda gross.