Implement caching for Annotations (#191)
Currently our caching setup assumes the underlying method takes exactly 1 argument, and that argument is a Pydantic object. Annotation takes a list of Pydantic objects, so I created a wrapper to hold that list. Also we already require all annotators to return a Pydantic object, so this is guaranteed to work with all annotators.

I tested this manually and it works. It is also exercised by `expensive-tests` via demo_04, and I verified that works as well.