Add link functions and view embed flag (#131)
* Add link functions and view embed flag

This allows for running `--view-embed` to generate a "emdedded view"
html. The main thing this is currently doing is simply not appending
`.html` to links. There will be more changes needed and the generic
term of `view-embed` vs doing `remove-extention` was in anticipation
of those other changes

This option defaults to false so users of the tool that are running
their own reports will not need to be aware of it as the default will
simply create the local html test report output.

new help text:

```
Usage: run.py benchmark [OPTIONS]

  run the standard benchmark

Options:
  -o, --output-dir DIRECTORY
  -m, --max-instances INTEGER
  --debug
  --web-only
  --view-embed                 Render the HTML to be emdedded in another view
  --help                       Show this message and exit.
```