Add dataurl for stylesheet on local build
* Add dataurl for stylesheet on local build
* Do not use dataurl for emdedded view

Use case: A someone that runs a benchmark, I want to be able to share an individual generated HTML file and have all the styling still applied.

Currently, if a single HTML file is shared with someone else via email or other mechanism, the href to the style.css file will not resolve. Because of this, they will not see the styling applied

This was lost in a merge/rebase and is #136 reapplied as that commit is not in the current history.