Use StringIO for csv-formatting (#235)
Right now, `make_csv()` does not properly format items that contain a CSV delimiter (e.g. ["a, b"] will be formatted and then read as a row with 2 items ["a", "b"]). 
This updates the method to use `io.StringIO()` instead, which is necessary to test Tests with prompts whose hazards contain commas (e.g. _"Suicide, self-harm and eating disorders"_).