Add a better error message for TogetherAI 400 errors (#210)
Under certain weird cases, TogetherAI will return a 400 error with important details in the HTTP body. This makes it throw something that includes the content. Not sure if there's a better exception to throw than Exception here; I rummaged a bit and the only thing I found was the together response_status_exception function, which just throws Exception.