Add test content and `content_root` property
* Add test content and `content_root` property to `StaticSiteGenerator`
  class
Allows for tests that test for whether certain components are visible in
the HTML while not requiring updates of those tests when updates are
made in the `content` folder. This is because these may be edited by
users that are not as familiar with the code base and we do not want to
make them maintain those tests as well.