Unify on "class_name" as the serialized fully qualified class name. (#270)
Before this PR, TypedData called it class_name, but InitializationRecord and SerializedSecret called it "qual_name". Since that json is likely to be read by users, I think "class_name" is more clear about what it is.

This breaks our ability to read JSON for those two objects written before this PR, but as I don't think we do that yet this should be fine.