Fix issue where BBQ tried to do zero test items.
In the docs it suggests the following command:
```
poetry run coffee benchmark -m 10
```
There are 11 BBQ instances, so this would cause `items` to be zero. Due to a [bug in newhelm](https://github.com/mlcommons/newhelm/pull/281) that would cause all 1000 BBQ instances to run for every BBQ test. Fixing the logic here so we get the desired 1 item per BBQ test.