Cache response json instead of pickle. (#341)
Pickle can execute arbitrary code, so previously sharing a cache file is a security issue. This also allows the cache responses to be human legible/queriable in SQL, which is a small plus I guess.

Rerunning the experiment from #325 shows this increase size a little bit (72k instead of 68k), but that probably isn't meaningful. If we really care about size, we could compress the json.