Remove echo_prompt as a SUTOption. (#339)
NewHELM inherited this from HELM. From what I can tell, echo_prompt's intent is to do language modeling. IMO that that is its own kind of SUT capability, and one not potentially supported by many high profile SUTs (e.g. OpenAIChat). I don't think its modeled right, and instead of trying to fix it, I figure we can remove it for now and reintroduce a better version when/if we need it.